tests/wpt/tests/websockets/interfaces/WebSocket/url/006.html

<!doctype html>
<title>WebSockets: defineProperty setter for url</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function() {
  window.setter_ran = false;
    Object.defineProperty(WebSocket.prototype, 'url', {
        set: function(v) { window[v] = true; }
    });
    var ws = new WebSocket('ws://example.invalid/');
    ws.url = 'setter_ran';
    assert_true(setter_ran);
});
</script>
